+++ stories/src/compat/buffer-geometry-utils.ts
@@ -1,0 +1,4 @@
+// @ts-ignore Three's examples package does not expose declarations for relative node_modules imports.
+export * from '../../../node_modules/three/examples/jsm/utils/BufferGeometryUtils.js';
+// @ts-ignore Three renamed mergeBufferGeometries to mergeGeometries in newer releases.
+export { mergeGeometries as mergeBufferGeometries } from '../../../node_modules/three/examples/jsm/utils/BufferGeometryUtils.js';