+++ stories/src/sunlight/sunlight.stories.tsx
@@ -1,4 +1,4 @@
-import { Billboard, Line, Plane, Ring, Sphere, useHelper } from "@react-three/drei";
+import { Billboard, Line, Plane, Ring, Sphere, useHelper } from "@react-three/drei/legacy";
 import { useFrame } from "@react-three/fiber";
 import { useControls } from "leva";
 import { RefObject, memo, useEffect, useMemo, useRef } from "react";
@@ -118,7 +118,7 @@ function Sun({ latitude, longitude }: { longitude: number, latitude: number }) {
   </>
 }
 
-function CamHelper({ camera }: { camera: RefObject<OrthographicCamera> }) {
+function CamHelper({ camera }: { camera: RefObject<OrthographicCamera | null> }) {
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
   useHelper(camera as any, CameraHelper);